Udated contactController.js, contactModel.js, contactRoutes.js so the auhtenticated user can only access it's contact for create read update and delete from the database.

+++ controller/contactController.js
@@ -1,18 +1,10 @@
 const asyncHandler = require("express-async-handler");
 const Contact = require('../models/contactModel');
 
-//@desc Get all contacts
-//@route GET /api/contacts
-//@access Public
-const getContacts = asyncHandler (async (req,res)=>{
-    const contacts = await Contact.find();
-    res.status(200).json( contacts )
-});
-
 
 //@desc Create contact
 //@route POST /api/contacts/
-//@access Public
+//@access Private
 const createContact = asyncHandler (async (req,res)=>{
     console.log("The request body is", req.body);
     const {name, email, phone} = req.body;
@@ -20,34 +12,52 @@ const createContact = asyncHandler (async (req,res)=>{
         res.status(400);
         throw new Error("All fields are necessary!!");
     }
+
     const contact = await Contact.create({
-        name, email, phone
+        user_id: req.user.id, name, email, phone
     })
-    res.status(201).json(contact);
+    res.status(201).json({message: "Created Contact",user: req.user.username, contact: contact});
+});
+
+
+//@desc Get all contacts
+//@route GET /api/contacts
+//@access Private
+const getContacts = asyncHandler (async (req,res)=>{
+    const contacts = await Contact.find({user_id: req.user.id});
+    res.status(200).json({user: req.user.username, contacts: contacts})
 });
 
 
 //@desc Get contact
 //@route GET /api/contacts/:id
-//@access Public
+//@access Private
 const getContact = asyncHandler (async (req,res)=>{
     const contact = await Contact.findById(req.params.id);
     if(!contact){
         res.status(404);
         throw new Error("Contact not found.");
     }
-    res.status(200).json(contact);
+    if(contact.user_id.toString() !== req.user.id){
+        res.status(403);
+        throw new Error("Your are unathorized to access.")
+    }
+    res.status(200).json({user: req.user.username, contacts: contact});
 });
 
 
 //@desc Update the contact
 //@route PUT /api/contacts
-//@access Public
+//@access Private
 const updateContact = asyncHandler (async (req,res)=>{
     const contact = await Contact.findById(req.params.id);
     if(!contact){
         res.status(404);
         throw new Error("Contact not found.");
+    }
+    if(contact.user_id.toString() !== req.user.id){
+        res.status(403);
+        throw new Error("Your are unathorized to access.")
     }
 
     const updatedContact = await Contact.findByIdAndUpdate(
@@ -55,23 +65,27 @@ const updateContact = asyncHandler (async (req,res)=>{
         req.body,
         {new: true}
     )
-    res.status(200).json(updatedContact);
+    res.status(200).json({message: "Updated Contact",user: req.user.username, contact: updatedContact});
 });
 
 
 //@desc Delete contact
 //@route DELETE /api/contacts
-//@access Public
+//@access Private
 const deleteContact = asyncHandler (async (req,res)=>{
     const contact = await Contact.findById(req.params.id);
     if(!contact){
         res.status(404);
         throw new Error("Contact not found.");
     }
+    if(contact.user_id.toString() !== req.user.id){
+        res.status(403);
+        throw new Error("Your are unathorized to delete other user contact.")
+    }
     // await contact.remove();
     await Contact.findByIdAndRemove(req.params.id);
     // await Contact.remove();
-    res.status(200).json(contact);
+    res.status(200).json({message: "Deleted Contact",user: req.user.username, contact: contact});
 });
 
 
@@ -81,4 +95,4 @@ module.exports = {
     updateContact,
     createContact,
     deleteContact
-}
+};
